+++ std/node/global.ts
@@ -1,4 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
+import process from "./process.ts";
+import { Buffer as buffer } from "./buffer.ts";
 
 Object.defineProperty(globalThis, Symbol.toStringTag, {
   value: "global",
@@ -9,5 +11,28 @@ Object.defineProperty(globalThis, Symbol.toStringTag, {
 
 // deno-lint-ignore no-explicit-any
 (globalThis as any)["global"] = globalThis;
+
+// Define the type for the global declration
+type Process = typeof process;
+type Buffer = typeof buffer;
+
+Object.defineProperty(globalThis, "process", {
+  value: process,
+  enumerable: false,
+  writable: true,
+  configurable: true,
+});
+
+declare global {
+  const process: Process;
+  const Buffer: Buffer;
+}
+
+Object.defineProperty(globalThis, "Buffer", {
+  value: buffer,
+  enumerable: false,
+  writable: true,
+  configurable: true,
+});
 
 export {};
